Struct uanft::uanft::AppData

pub struct AppData {
    pub token_instance: <ILOCKmvpRef as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type,
    pub operator: <AccountID as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type,
    pub _reserved: <Option<()> as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type,
}

Fields
token_instance: <ILOCKmvpRef as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type

This is PSP22 token contract that this uanft application connects to via socket.

This is used for self-minting purposes, which means non-owner can mint in excange for PSP22 token (ILOCK in this case) without needing to rely on a transaction relay server off-chain.

operator: <AccountID as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type

This is address that manages this uanft contract and receives ILOCK (or other) PSP22 token for self-mint transactions.

_reserved: <Option<()> as AutoStorableHint<ManualKey<{ _ }, ManualKey<APP_KEY>>>>::Type

This is to expand storage related to this uanft application functionality.