Struct uanft::uanft::CallBuilder

#[repr(transparent)]
pub struct CallBuilder {
    account_id: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId,
}

The ink! smart contract’s call builder.

Implements the underlying on-chain calling of the ink! smart contract messages and trait implementations in a type safe way.

Fields
account_id: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId
Implementations

impl CallBuilder

pub fn order_multisigtx( &mut self, __ink_binding_0: String ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, EmptyArgumentList>>>, Set<ReturnType<OtherResult<()>>>>

Function to order multisigtx transaction.

pub fn sign_multisigtx( &mut self, __ink_binding_0: String ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, EmptyArgumentList>>>, Set<ReturnType<OtherResult<()>>>>

A multisigtx signer calls this to sign.

pub fn add_signatory( &mut self, __ink_binding_0: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: String ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

This adds a signatory from the list of permitted signatories.

pub fn remove_signatory( &mut self, __ink_binding_0: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: String ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

This removes a signatory from the list of permitted signatories.

pub fn change_threshold( &mut self, __ink_binding_0: u16, __ink_binding_1: String ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<u16>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

This changes signer threshold for approving multisigtx.

pub fn change_multisigtxtimelimit( &mut self, __ink_binding_0: <<Psp34Nft as ContractEnv>::Env as Environment>::Timestamp, __ink_binding_1: String ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::Timestamp>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

This modifies timelimit for a multisig transaction.

pub fn threshold( &self ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<u16>>>

This gets the current signature threshold for multisigtx.

pub fn multisigtimelimit( &self ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<<<Psp34Nft as ContractEnv>::Env as Environment>::Timestamp>>>

This gets the current timelimit for signatories to sign multisigtx.

pub fn signatories( &mut self ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<OtherResult<Vec<AccountID>>>>>

This gets a list of current accounts permitted to sign multisigtx.

pub fn signatory_count( &self ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<u8>>>

This gets number of signatories permitted to sign multisigtx.

pub fn signature_count( &self ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<u8>>>

This gets current number of signatures for multisigtx.

pub fn check_signatures( &mut self ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<OtherResult<Vec<Signature>>>>>

This gets a list of all signers so far on a multisigtx.

pub fn pause( &mut self ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<OtherResult<()>>>>

Function pauses contract.

Any signatory may call.

pub fn unpause( &mut self, __ink_binding_0: String ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, EmptyArgumentList>>>, Set<ReturnType<OtherResult<()>>>>

Function unpauses contract.

pub fn mint( &mut self ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<Result<(), Error>>>>

This generic mint function is for Art Zero interface.

pub fn mint_to( &mut self, __ink_binding_0: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>, Set<ReturnType<Result<(), Error>>>>

This mints a universal access nft by Interlock Network to specific recipient.

pub fn self_mint( &mut self, __ink_binding_0: <<Psp34Nft as ContractEnv>::Env as Environment>::Balance ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::Balance>, EmptyArgumentList>>>, Set<ReturnType<Result<(), Error>>>>

This mints a universal access nft to caller’s self at token_price in terms of PSP22 token.

pub fn mint_with_attributes( &mut self, __ink_binding_0: Vec<(String, String)> ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<Vec<(String, String)>>, EmptyArgumentList>>>, Set<ReturnType<Result<(), Error>>>>

This is a mint function for Art Zero interface.

pub fn create_socket( &mut self ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<Result<(), OtherError>>>>

This registers this universal access nft contract with ILOCK PSP22 token contract to allow self-minting.

Only contract owner may create a socket between this contract and the ILOCK PSP22 token.

pub fn call_socket( &mut self, __ink_binding_0: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: <<Psp34Nft as ContractEnv>::Env as Environment>::Balance, __ink_binding_2: Vec<u8> ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<Vec<u8>>, ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::Balance>, ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>>, Set<ReturnType<Result<(), OtherError>>>>

This makes call through universal access nft socket to ILOCK PSP22 token contract on port 0 or port 1, depending on this contract’s configuration and affiliation with Interlock Network.

(Ie, transfer token from recipient to contract owner within PSP22 contract.)

Only operator may call.

pub fn get_token_price( &self ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<EmptyArgumentList>>, Set<ReturnType<<<Psp34Nft as ContractEnv>::Env as Environment>::Balance>>>

Retrieve the current price of universal access nft self-minting.

pub fn set_token_price( &mut self, __ink_binding_0: <<Psp34Nft as ContractEnv>::Env as Environment>::Balance ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::Balance>, EmptyArgumentList>>>, Set<ReturnType<Result<(), PSP34Error>>>>

Owner may change the price that self-minter must pay for universal access nft.

pub fn get_collection( &self, __ink_binding_0: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>, Set<ReturnType<Result<Vec<Id>, PSP34Error>>>>

Get collection of nfts held by particular address.

pub fn transfer_ownership( &mut self, __ink_binding_0: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId, __ink_binding_1: String ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

pub fn update_contract( &mut self, __ink_binding_0: [u8; 32], __ink_binding_1: String ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<String>, ArgumentList<Argument<[u8; 32]>, EmptyArgumentList>>>>, Set<ReturnType<OtherResult<()>>>>

Modifies the code which is used to execute calls to this contract address.

This upgrades the token contract logic while using old state.

pub fn contract_hash( &self, __ink_binding_0: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId ) -> CallBuilder<<Psp34Nft as ContractEnv>::Env, Set<Call<<Psp34Nft as ContractEnv>::Env>>, Set<ExecutionInput<ArgumentList<Argument<<<Psp34Nft as ContractEnv>::Env as Environment>::AccountId>, EmptyArgumentList>>>, Set<ReturnType<Result<<<Psp34Nft as ContractEnv>::Env as Environment>::Hash, PSP34Error>>>>

This is a testing helper for port applications.

Returns code hash for contract running e2e test (for port application verification).

Otherwise, each attempt to hardcode hash in application contract changes the hash for that run.