Struct uanft::uanft::Psp34Nft

pub struct Psp34Nft {

    psp34: <Data<Balances> as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    metadata: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    ownable: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    pausable: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    access: <AccessData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    app: <AppData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    multisig: <MultisigData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    last_token_id: <u64 as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    attribute_count: <u32 as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    attribute_names: <Mapping<u32, Vec<u8>> as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    locked_tokens: <Mapping<Id, bool> as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    locked_token_count: <u64 as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,
    is_attribute: <Mapping<String, bool> as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type,

}

Main contract storage.

Fields
psp34: <Data<Balances> as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush PSP34 storage fields.

metadata: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush metadata extension storage fields.

ownable: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush ownable extension storage fields.

pausable: <Data as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Openbrush pausable extensios.

access: <AccessData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Universal access NFT storage fields.

app: <AppData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Storage fields related to the UANFT as an application for the ILOCK PSP22 contract.

multisig: <MultisigData as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

ILOCK multisig contract info.

last_token_id: <u64 as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type

Art zero storage fields.

attribute_count: <u32 as AutoStorableHint<ManualKey<{ _ }, ()>>>::Typeattribute_names: <Mapping<u32, Vec<u8>> as AutoStorableHint<ManualKey<{ _ }, ()>>>::Typelocked_tokens: <Mapping<Id, bool> as AutoStorableHint<ManualKey<{ _ }, ()>>>::Typelocked_token_count: <u64 as AutoStorableHint<ManualKey<{ _ }, ()>>>::Typeis_attribute: <Mapping<String, bool> as AutoStorableHint<ManualKey<{ _ }, ()>>>::Type
Implementations

impl Psp34Nft

pub fn new( name: String, symbol: String, class: String, cap: u64, price: <<Psp34Nft as ContractEnv>::Env as Environment>::Balance, token_address: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId, timelimit: <<Psp34Nft as ContractEnv>::Env as Environment>::Timestamp, signatory_2: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId, signatory_3: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId ) -> Self

UANFT contract constructor.

pub fn order_multisigtx(&mut self, function: String) -> OtherResult<()>

Function to order multisigtx transaction.

pub fn sign_multisigtx(&mut self, function: String) -> OtherResult<()>

A multisigtx signer calls this to sign.

pub fn add_signatory( &mut self, signatory: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId, function: String ) -> OtherResult<()>

This adds a signatory from the list of permitted signatories.

pub fn remove_signatory( &mut self, signatory: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId, function: String ) -> OtherResult<()>

This removes a signatory from the list of permitted signatories.

pub fn change_threshold( &mut self, threshold: u16, function: String ) -> OtherResult<()>

This changes signer threshold for approving multisigtx.

pub fn change_multisigtxtimelimit( &mut self, timelimit: <<Psp34Nft as ContractEnv>::Env as Environment>::Timestamp, function: String ) -> OtherResult<()>

This modifies timelimit for a multisig transaction.

pub fn threshold(&self) -> u16

This gets the current signature threshold for multisigtx.

pub fn multisigtimelimit( &self ) -> <<Psp34Nft as ContractEnv>::Env as Environment>::Timestamp

This gets the current timelimit for signatories to sign multisigtx.

pub fn signatories(&mut self) -> OtherResult<Vec<AccountID>>

This gets a list of current accounts permitted to sign multisigtx.

pub fn signatory_count(&self) -> u8

This gets number of signatories permitted to sign multisigtx.

pub fn signature_count(&self) -> u8

This gets current number of signatures for multisigtx.

pub fn check_signatures(&mut self) -> OtherResult<Vec<Signature>>

This gets a list of all signers so far on a multisigtx.

pub fn pause(&mut self) -> OtherResult<()>

Function pauses contract.

Any signatory may call.

pub fn unpause(&mut self, function: String) -> OtherResult<()>

Function unpauses contract.

pub fn mint(&mut self) -> Result<(), Error>

This generic mint function is for Art Zero interface.

pub fn mint_to( &mut self, recipient: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId ) -> Result<(), Error>

This mints a universal access nft by Interlock Network to specific recipient.

pub fn self_mint( &mut self, price: <<Psp34Nft as ContractEnv>::Env as Environment>::Balance ) -> Result<(), Error>

This mints a universal access nft to caller’s self at token_price in terms of PSP22 token.

pub fn mint_with_attributes( &mut self, metadata: Vec<(String, String)> ) -> Result<(), Error>

This is a mint function for Art Zero interface.

pub fn create_socket(&mut self) -> Result<(), OtherError>

This registers this universal access nft contract with ILOCK PSP22 token contract to allow self-minting.

Only contract owner may create a socket between this contract and the ILOCK PSP22 token.

pub fn call_socket( &mut self, address: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId, amount: <<Psp34Nft as ContractEnv>::Env as Environment>::Balance, data: Vec<u8> ) -> Result<(), OtherError>

This makes call through universal access nft socket to ILOCK PSP22 token contract on port 0 or port 1, depending on this contract’s configuration and affiliation with Interlock Network.

(Ie, transfer token from recipient to contract owner within PSP22 contract.)

Only operator may call.

pub fn get_token_price( &self ) -> <<Psp34Nft as ContractEnv>::Env as Environment>::Balance

Retrieve the current price of universal access nft self-minting.

pub fn set_token_price( &mut self, price: <<Psp34Nft as ContractEnv>::Env as Environment>::Balance ) -> Result<(), PSP34Error>

Owner may change the price that self-minter must pay for universal access nft.

pub fn get_collection( &self, address: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId ) -> Result<Vec<Id>, PSP34Error>

Get collection of nfts held by particular address.

pub fn transfer_ownership( &mut self, newowner: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId, function: String ) -> OtherResult<()>

pub fn update_contract( &mut self, code_hash: [u8; 32], function: String ) -> OtherResult<()>

Modifies the code which is used to execute calls to this contract address.

This upgrades the token contract logic while using old state.

pub fn contract_hash( &self, application: <<Psp34Nft as ContractEnv>::Env as Environment>::AccountId ) -> Result<<<Psp34Nft as ContractEnv>::Env as Environment>::Hash, PSP34Error>

This is a testing helper for port applications.

Returns code hash for contract running e2e test (for port application verification).

Otherwise, each attempt to hardcode hash in application contract changes the hash for that run.

pub fn emit_event<EE: EmitEvent<Self>>(emitter: EE, event: Event)

Function for internal _emit_event implementations.

pub fn check_multisig(&self, function: String) -> OtherResult<()>

Helper function for checking signature count

pub fn add_attribute_name(&mut self, attribute_input: &Vec<u8>)

Art Zero helper function.