Module uanft::uanft

Modules

psp34traits_external

Structs

AccessData

AccountID

This is a type wrapper to implement Default method on AccountId type. Ink 4 stable eliminated AccountId Default (which was zero address, that has known private key) …we only really need this because Openbrush contract relies on deriving Default for contract storage, and our AccesData struct contains AccountId.

AppData

Approval

Specify approval event.

CallBuilder

The ink! smart contract’s call builder.

Check 🔒

Psp34Nft

Main contract storage.

Psp34NftRef

Main contract storage.

Transfer

Specify transfer event.

Enums

Error

__ink_ConstructorDecoder

__ink_EventBase

__ink_MessageDecoder

__ink_UndefinedAmountOfTopics

Constants

ACCESS_KEY

This is upgradable storage for the access features for this universal access nft contract.

APP_KEY

This is upgradable storage for the features that allow this universal access nft contract to connect as an application to the ILOCK (or other) PSP22 contract the application socket abstraction.

ARR_LEN 🔒

ARR_LEN 🔒

CONCATP_NHPMWYD3NJA 🔒

CONCATP_NHPMWYD3NJA 🔒

CONCAT_ARR 🔒

CONCAT_ARR 🔒

CONCAT_STR 🔒

CONCAT_STR 🔒

CONSTRUCTOR_0 🔒

MESSAGE_0 🔒

MESSAGE_1 🔒

MESSAGE_2 🔒

MESSAGE_3 🔒

MESSAGE_4 🔒

MESSAGE_5 🔒

MESSAGE_6 🔒

MESSAGE_7 🔒

MESSAGE_8 🔒

MESSAGE_9 🔒

MESSAGE_10 🔒

MESSAGE_11 🔒

MESSAGE_12 🔒

MESSAGE_13 🔒

MESSAGE_14 🔒

MESSAGE_15 🔒

MESSAGE_16 🔒

MESSAGE_17 🔒

MESSAGE_18 🔒

MESSAGE_19 🔒

MESSAGE_20 🔒

MESSAGE_21 🔒

MESSAGE_22 🔒

MESSAGE_23 🔒

MESSAGE_24 🔒

MESSAGE_25 🔒

MESSAGE_26 🔒

MESSAGE_27 🔒

MESSAGE_28 🔒

MESSAGE_29 🔒

MESSAGE_30 🔒

MESSAGE_31 🔒

MESSAGE_32 🔒

MESSAGE_33 🔒

MESSAGE_34 🔒

MESSAGE_35 🔒

MESSAGE_36 🔒

MESSAGE_37 🔒

MESSAGE_38 🔒

MESSAGE_39 🔒

MESSAGE_40 🔒

PORT

This is the port number for this type of uanft application socket connections to ILOCK (or other) PSP22 token contract.

PORT 0 designates uanft contracts owned by Interlock Network.

This port is locked by default (only Interlock Network may connect this uanft contract via socket to ILOCK PSP22 contrac.

Traits

Psp34Traits

Art Zero trait definitions.

UnionPSP34WrapperPSP34MetadataWrapper

Functions

__ink_generate_metadata

call 🔒

deploy 🔒

push_contract 🔒

Type Definitions

AccountId 🔒

Balance 🔒

BlockNumber 🔒

Environment 🔒

Event

Needed for Openbrush internal event emission implementations.

Hash 🔒

Psp34Ref

Timestamp 🔒